feat: add manifests page and sub-navbar to registry detail
Restructures the repository detail page into tabs (Tags / Manifests)
using a shared layout. The existing /registry/detail URL redirects to
/registry/detail/tags for backward compatibility. The new Manifests tab
lists all manifests with their digest and creation time, and allows
deleting a manifest via registry/deleteManifest.
